Implement lines for debugging and simplify the search for previous solutions
Some comment lines should be uncommented to debug
The search for previous solutions is now implemented more simply. The solutions with
the efArray are inserted in a list. To know if a previous efArray produce solutions
that can satisfy the current efArray we have to search in the list, time O(n)