Hot fix in SAUGMECON non-recursive: avoid false infeasibility; add tests
In the multi-objective knapsack, the constraint PropKnapsackKatriel01 can make the solver to launch infeasibility for regions of the objective space which are feasible.
The fastest workaround is to simulate a restart.
In the multi-objective knapsack, the constraint
PropKnapsackKatriel01 can incorrectly mark regions of the objective
space as infeasible even though they are feasible. This led the solver
to report infeasibility in some algorithms.

Fix
Simulate a solver restart.

Tests
Add tests for several multi-objective instances.